+++ axes/decorators.py
@@ -2,11 +2,10 @@ from functools import wraps
 
 from axes.handlers.proxy import AxesProxyHandler
 from axes.helpers import get_lockout_response
-from axes.request import AxesHttpRequest
 
 
 def axes_dispatch(func):
-    def inner(request: AxesHttpRequest, *args, **kwargs):
+    def inner(request, *args, **kwargs):
         if AxesProxyHandler.is_allowed(request):
             return func(request, *args, **kwargs)