+++ axes/handlers/dummy.py
@@ -1,5 +1,3 @@
-from django.http import HttpRequest
-
 from axes.handlers.base import AxesHandler
 
 
@@ -8,5 +6,5 @@ class AxesDummyHandler(AxesHandler):  # pylint: disable=unused-argument
     Signal handler implementation that does nothing and can be used to disable signal processing.
     """
 
-    def is_allowed(self, request: HttpRequest, credentials: dict = None) -> bool:
+    def is_allowed(self, request, credentials: dict = None) -> bool:
         return True