+++ axes/middleware.py
@@ -1,18 +1,6 @@
 from typing import Callable
 
-from django.http import HttpRequest
-from django.utils.timezone import now
-
-from axes.exceptions import AxesSignalPermissionDenied
-from axes.helpers import (
-    get_client_ip_address,
-    get_client_user_agent,
-    get_client_path_info,
-    get_client_http_accept,
-    get_lockout_response,
-    toggleable,
-)
-from axes.request import AxesHttpRequest
+from axes.helpers import get_lockout_response
 
 
 class AxesMiddleware:
@@ -42,35 +30,10 @@ class AxesMiddleware:
     def __init__(self, get_response: Callable):
         self.get_response = get_response
 
-    def __call__(self, request: HttpRequest):
-        self.update_request(request)
-        return self.get_response(request)
+    def __call__(self, request):
+        response = self.get_response(request)
 
-    @toggleable
-    def update_request(self, request: HttpRequest):
-        """
-        Construct an ``AxesHttpRequest`` from the given ``HttpRequest``
-        by updating the request with necessary attempt tracking attributes.
+        if getattr(request, 'axes_locked_out', None):
+            response = get_lockout_response(request)  # type: ignore
 
-        This method is called by the middleware class ``__call__`` method
-        when iterating over the middleware stack.
-        """
-
-        request.axes_attempt_time = now()
-        request.axes_ip_address = get_client_ip_address(request)
-        request.axes_user_agent = get_client_user_agent(request)
-        request.axes_path_info = get_client_path_info(request)
-        request.axes_http_accept = get_client_http_accept(request)
-
-    @toggleable
-    def process_exception(self, request: AxesHttpRequest, exception):  # pylint: disable=inconsistent-return-statements
-        """
-        Handle exceptions raised by the Axes signal handler class when requests fail checks.
-
-        Note that only ``AxesSignalPermissionDenied`` is handled by this middleware class.
-
-        :return: Configured ``HttpResponse`` for failed authentication attempts and lockouts.
-        """
-
-        if isinstance(exception, AxesSignalPermissionDenied):
-            return get_lockout_response(request)
+        return response