+++ axes/tests/test_utils.py
@@ -2,7 +2,7 @@ from datetime import timedelta
 from hashlib import md5
 from unittest.mock import patch
 
-from django.http import JsonResponse, HttpResponseRedirect, HttpResponse
+from django.http import JsonResponse, HttpResponseRedirect, HttpResponse, HttpRequest
 from django.test import override_settings, RequestFactory
 
 from axes import get_version
@@ -22,7 +22,6 @@ from axes.helpers import (
     is_ip_address_in_whitelist,
     is_client_method_whitelisted,
 )
-from axes.request import AxesHttpRequest
 
 
 class VersionTestCase(AxesTestCase):
@@ -367,7 +366,7 @@ class UsernameTestCase(AxesTestCase):
     def test_default_get_client_username(self):
         expected = 'test-username'
 
-        request = AxesHttpRequest()
+        request = HttpRequest()
         request.POST['username'] = expected
 
         actual = get_client_username(request)
@@ -379,7 +378,7 @@ class UsernameTestCase(AxesTestCase):
         expected = 'test-username'
         expected_in_credentials = 'test-credentials-username'
 
-        request = AxesHttpRequest()
+        request = HttpRequest()
         request.POST['username'] = expected
         credentials = {
             'username': expected_in_credentials
@@ -399,7 +398,7 @@ class UsernameTestCase(AxesTestCase):
         expected = 'prefixed-' + provided
         provided_in_credentials = 'test-credentials-username'
 
-        request = AxesHttpRequest()
+        request = HttpRequest()
         request.POST['username'] = provided
         credentials = {'username': provided_in_credentials}
 
@@ -417,7 +416,7 @@ class UsernameTestCase(AxesTestCase):
         provided_in_credentials = 'test-credentials-username'
         expected_in_credentials = 'prefixed-' + provided_in_credentials
 
-        request = AxesHttpRequest()
+        request = HttpRequest()
         request.POST['username'] = provided
         credentials = {'username': provided_in_credentials}
 
@@ -427,38 +426,38 @@ class UsernameTestCase(AxesTestCase):
 
     @override_settings(AXES_USERNAME_CALLABLE=lambda request, credentials: 'example')  # pragma: no cover
     def test_get_client_username(self):
-        self.assertEqual(get_client_username(AxesHttpRequest(), {}), 'example')
+        self.assertEqual(get_client_username(HttpRequest(), {}), 'example')
 
     @override_settings(AXES_USERNAME_CALLABLE=lambda request: None)  # pragma: no cover
     def test_get_client_username_invalid_callable_too_few_arguments(self):
         with self.assertRaises(TypeError):
-            get_client_username(AxesHttpRequest(), {})
+            get_client_username(HttpRequest(), {})
 
     @override_settings(AXES_USERNAME_CALLABLE=lambda request, credentials, extra: None)  # pragma: no cover
     def test_get_client_username_invalid_callable_too_many_arguments(self):
         with self.assertRaises(TypeError):
-            get_client_username(AxesHttpRequest(), {})
+            get_client_username(HttpRequest(), {})
 
     @override_settings(AXES_USERNAME_CALLABLE=True)
     def test_get_client_username_not_callable(self):
         with self.assertRaises(TypeError):
-            get_client_username(AxesHttpRequest(), {})
+            get_client_username(HttpRequest(), {})
 
     @override_settings(AXES_USERNAME_CALLABLE='axes.tests.test_utils.get_username')
     def test_get_client_username_str(self):
         self.assertEqual(
-            get_client_username(AxesHttpRequest(), {}),
+            get_client_username(HttpRequest(), {}),
             'username',
         )
 
 
-def get_username(request: AxesHttpRequest, credentials: dict) -> str:
+def get_username(request, credentials: dict) -> str:
     return 'username'
 
 
 class IPWhitelistTestCase(AxesTestCase):
     def setUp(self):
-        self.request = AxesHttpRequest()
+        self.request = HttpRequest()
         self.request.method = 'POST'
         self.request.META['REMOTE_ADDR'] = '127.0.0.1'
         self.request.axes_ip_address = '127.0.0.1'
@@ -517,7 +516,7 @@ class IPWhitelistTestCase(AxesTestCase):
 
 class MethodWhitelistTestCase(AxesTestCase):
     def setUp(self):
-        self.request = AxesHttpRequest()
+        self.request = HttpRequest()
         self.request.method = 'GET'
 
     @override_settings(AXES_NEVER_LOCKOUT_GET=True)
@@ -531,7 +530,7 @@ class MethodWhitelistTestCase(AxesTestCase):
 
 class LockoutResponseTestCase(AxesTestCase):
     def setUp(self):
-        self.request = AxesHttpRequest()
+        self.request = HttpRequest()
 
     @override_settings(AXES_COOLOFF_TIME=42)
     def test_get_lockout_response_cool_off(self):